Inject NavigationManager and add interactive button
In `App.razor`, the `NavigationManager` is now injected to dynamically set the `_baseUrl` variable, which is used to set the base href in the HTML head. A new interactive button has been added in `InteractiveButton.razor` that updates a feedback message with the current local time. The navigation menu in `NavMenu.razor` has been restructured, removing unnecessary div elements and adding a new navigation link to the "Base Url" page. A new page has been created in `BaseUrl.razor` to display various base URLs and an interactive button.